# Don't reset computed properties if already empty